# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Seal Platform API
  version: 2.0.0
  description: 印章平台 V2 已实现 HTTP 契约。除健康检查、同步预览、资源目录、支付回调和 CSP 报告外均使用 HttpOnly Cookie 会话。
servers:
  - url: /api/v1
...
paths:
...
  /seals/render:
    post:
      summary: 同步渲染 SVG
      requestBody: {$ref: '#/components/requestBodies/SealConfig'}
      responses:
        '200': {description: SVG, content: {image/svg+xml: {schema: {type: string}}}}
        '422': {$ref: '#/components/responses/Error'}
...
  /generations:
    get:
      tags: [Generations]
      summary: 本人生成历史
      responses:
        '200': {description: 任务列表, content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Generation'}}}}}}}
    post:
      tags: [Generations]
      summary: 创建幂等高清任务；服务端重算尺寸与水印
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [config, format], properties: {format: {type: string, enum: [svg, png]}, config: {$ref: './seal-config.schema.json'}}}
      responses:
        '202': {description: 已排队, content: {application/json: {schema: {$ref: '#/components/schemas/Generation'}}}}
        '429': {$ref: '#/components/responses/Error'}
...
  /generations/{id}/download-token:
    post:
      tags: [Generations]
      summary: 签发 120 秒、绑定用户与任务的单次令牌
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: 下载令牌, content: {application/json: {schema: {type: object, required: [token, expiresAt, downloadUrl], properties: {token: {type: string}, expiresAt: {type: string, format: date-time}, downloadUrl: {type: string}}}}}}
        '403': {$ref: '#/components/responses/Error'}
...
components:
  parameters:
    ID: {in: path, name: id, required: true, schema: {type: string}}
  requestBodies:
...
    SealConfig:
      required: true
      content: {application/json: {schema: {$ref: './seal-config.schema.json'}}}
...
  responses:
    Error:
      description: 结构化错误
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
  schemas:
    Error: {type: object, properties: {error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}}}
...
    Generation: {type: object, required: [id, userId, config, rendererVersion, format, status, watermark, createdAt], properties: {id: {type: string}, userId: {type: string}, config: {$ref: './seal-config.schema.json'}, rendererVersion: {type: string}, format: {type: string, enum: [svg, png]}, status: {type: string, enum: [queued, rendering, succeeded, failed]}, watermark: {type: boolean}, accessRevoked: {type: boolean}, failureReason: {type: string}, createdAt: {type: string, format: date-time}, finishedAt: {type: string, format: date-time}}}
...